Attached front-end stuff to site, and created home
I attached everything in bower_components/ to the /assets/ path, and
I attached all the resources in public/ to the resources/ path. I made
some modifications to home.ejs to create a login panel and a nav-bar.
I also set up the general color scheme of the site. I will be using a
lot of orange.