Added color coding to visually identify quantity
If the item has more of an on hand quantity than what is needed then the
table entry is highlighted with green. If the item has less of an on hand
quantity than is needed then the table entry is highlighted with red.
Finally if the needed quantity and on hand quantity are the same then
the table entry is highlighted with yellow to indicate that it is running
out.